+++ tools/mpremote/mpremote/mip.py
@@ -16,7 +16,7 @@ _CHUNK_SIZE = 128
 
 
 # This implements os.makedirs(os.dirname(path))
-def _ensure_path_exists(pyb, path):
+def _ensure_path_exists(transport, path):
     import os
 
     split = path.split("/")
@@ -29,8 +29,8 @@ def _ensure_path_exists(pyb, path):
     prefix = ""
     for i in range(len(split) - 1):
         prefix += split[i]
-        if not pyb.fs_exists(prefix):
-            pyb.fs_mkdir(prefix)
+        if not transport.fs_exists(prefix):
+            transport.fs_mkdir(prefix)
         prefix += "/"
 
 
@@ -68,7 +68,7 @@ def _rewrite_url(url, branch=None):
     return url
 
 
-def _download_file(pyb, url, dest):
+def _download_file(transport, url, dest):
     try:
         with urllib.request.urlopen(url) as src:
             fd, path = tempfile.mkstemp()
@@ -76,8 +76,8 @@ def _download_file(pyb, url, dest):
                 print("Installing:", dest)
                 with os.fdopen(fd, "wb") as f:
                     _chunk(src, f.write, src.length)
-                _ensure_path_exists(pyb, dest)
-                pyb.fs_put(path, dest, progress_callback=show_progress_bar)
+                _ensure_path_exists(transport, dest)
+                transport.fs_put(path, dest, progress_callback=show_progress_bar)
             finally:
                 os.unlink(path)
     except urllib.error.HTTPError as e:
@@ -89,7 +89,7 @@ def _download_file(pyb, url, dest):
         raise CommandError(f"{e.reason} requesting {url}")
 
 
-def _install_json(pyb, package_json_url, index, target, version, mpy):
+def _install_json(transport, package_json_url, index, target, version, mpy):
     try:
         with urllib.request.urlopen(_rewrite_url(package_json_url, version)) as response:
             package_json = json.load(response)
@@ -103,15 +103,15 @@ def _install_json(pyb, package_json_url, index, target, version, mpy):
     for target_path, short_hash in package_json.get("hashes", ()):
         fs_target_path = target + "/" + target_path
         file_url = f"{index}/file/{short_hash[:2]}/{short_hash}"
-        _download_file(pyb, file_url, fs_target_path)
+        _download_file(transport, file_url, fs_target_path)
     for target_path, url in package_json.get("urls", ()):
         fs_target_path = target + "/" + target_path
-        _download_file(pyb, _rewrite_url(url, version), fs_target_path)
+        _download_file(transport, _rewrite_url(url, version), fs_target_path)
     for dep, dep_version in package_json.get("deps", ()):
-        _install_package(pyb, dep, index, target, dep_version, mpy)
+        _install_package(transport, dep, index, target, dep_version, mpy)
 
 
-def _install_package(pyb, package, index, target, version, mpy):
+def _install_package(transport, package, index, target, version, mpy):
     if (
         package.startswith("http://")
         or package.startswith("https://")
@@ -120,7 +120,7 @@ def _install_package(pyb, package, index, target, version, mpy):
         if package.endswith(".py") or package.endswith(".mpy"):
             print(f"Downloading {package} to {target}")
             _download_file(
-                pyb, _rewrite_url(package, version), target + "/" + package.rsplit("/")[-1]
+                transport, _rewrite_url(package, version), target + "/" + package.rsplit("/")[-1]
             )
             return
         else:
@@ -136,14 +136,15 @@ def _install_package(pyb, package, index, target, version, mpy):
 
         mpy_version = "py"
         if mpy:
-            pyb.exec("import sys")
+            transport.exec("import sys")
             mpy_version = (
-                int(pyb.eval("getattr(sys.implementation, '_mpy', 0) & 0xFF").decode()) or "py"
+                int(transport.eval("getattr(sys.implementation, '_mpy', 0) & 0xFF").decode())
+                or "py"
             )
 
         package = f"{index}/package/{mpy_version}/{package}/{version}.json"
 
-    _install_json(pyb, package, index, target, version, mpy)
+    _install_json(transport, package, index, target, version, mpy)
 
 
 def do_mip(state, args):
@@ -163,9 +164,9 @@ def do_mip(state, args):
                 args.index = _PACKAGE_INDEX
 
             if args.target is None:
-                state.pyb.exec("import sys")
+                state.transport.exec("import sys")
                 lib_paths = (
-                    state.pyb.eval("'\\n'.join(p for p in sys.path if p.endswith('/lib'))")
+                    state.transport.eval("'\\n'.join(p for p in sys.path if p.endswith('/lib'))")
                     .decode()
                     .split("\n")
                 )
@@ -181,7 +182,12 @@ def do_mip(state, args):
 
             try:
                 _install_package(
-                    state.pyb, package, args.index.rstrip("/"), args.target, version, args.mpy
+                    state.transport,
+                    package,
+                    args.index.rstrip("/"),
+                    args.target,
+                    version,
+                    args.mpy,
                 )
             except CommandError:
                 print("Package may be partially installed")
