+++ tools/mpremote/pyproject.toml
@@ -44,11 +44,5 @@ raw-options = { root = "../..", version_scheme = "post-release" }
 [tool.hatch.build]
 packages = ["mpremote"]
 
-# Also grab pyboard.py from /tools and add it to the package for both wheel and sdist.
-[tool.hatch.build.force-include]
-"../pyboard.py" = "mpremote/pyboard.py"
-
-# Workaround to allow `python -m build` to work.
 [tool.hatch.build.targets.sdist.force-include]
-"../pyboard.py" = "mpremote/pyboard.py"
 "requirements.txt" = "requirements.txt"
